Added PayPal integration, updated package-lock.json and package.json, modified several files in src/app and src/components
