+++ package.json
@@ -10,6 +10,7 @@
     "seed": "ts-node src/seed/seed-database.ts"
   },
   "dependencies": {
+    "@paypal/react-paypal-js": "^8.5.0",
     "@prisma/client": "^5.17.0",
     "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
